requirements.txt
websocket-client
tornado
git+https://github.com/grangier/python-goose.git
gevent
